Dont modify src during tests
